Move extractInlineImagesToAttachments() method from PlainMsgMailFormatter to Mime
