+++ extension/chrome/elements/compose-modules/compose-send-btn-popover-module.ts
@@ -29,7 +29,7 @@ export class ComposeSendBtnPopoverModule extends ViewModule<ComposeView> {
     for (const key of Object.keys(popoverItems)) {
       const popoverOpt = key as PopoverOpt;
       if (popoverOpt === 'richtext' && !this.view.debug && Catch.browser().name !== 'firefox') {
-        continue; // richtext not deployed to Chrome yet, for now only allow firefox (and also in automated tests which set debug===true)
+        // continue; // richtext not deployed to Chrome yet, for now only allow firefox (and also in automated tests which set debug===true)
       }
       const item = popoverItems[popoverOpt];
       const elem = $(`
